+++ apps/meteor/server/lib/oauth/addPassportCustomOAuth.ts
@@ -1,17 +1,12 @@
-import type { OAuthConfiguration, IUser } from '@rocket.chat/core-typings';
-import type { Request, Response } from 'express';
-import { Accounts } from 'meteor/accounts-base';
+import type { OAuthConfiguration } from '@rocket.chat/core-typings';
 import passport from 'passport';
 import type { DoneCallback, Profile } from 'passport';
 
+import { passportOAuthCallback } from './passportOAuthCallback';
 import { verifyFunction } from './verifyFunction';
 import { CustomOAuthStrategy } from '../../../app/custom-oauth/server/customOAuth';
 import { settings } from '../../../app/settings/server';
 import { oAuthRouter } from '../../configuration/configurePassport';
-
-interface IOAuthRequest extends Request {
-	user?: IUser;
-}
 
 export const addPassportCustomOAuth = (serviceName: string, config: Partial<OAuthConfiguration & { clientSecret: string }>) => {
 	passport.unuse(serviceName);
@@ -47,34 +42,6 @@ export const addPassportCustomOAuth = (serviceName: string, config: Partial<OAut
 	oAuthRouter.get(
 		`/oauth/${serviceName}/callback`,
 		passport.authenticate(serviceName, { failureRedirect: '/login', failureFlash: true, failWithError: true, keepSessionInfo: true }),
-		async (req: IOAuthRequest, res: Response) => {
-			const oAuthUser = req.user as IUser;
-
-			if (!oAuthUser) {
-				return res.redirect('/login');
-			}
-
-			const { loginClient } = req.session;
-
-			const stampedToken = Accounts._generateStampedLoginToken();
-			await Accounts._insertLoginToken(oAuthUser._id, stampedToken);
-
-			const redirectUrl = new URL(`/home`, siteUrl);
-
-			redirectUrl.searchParams.set('resumeToken', stampedToken.token);
-			redirectUrl.searchParams.set('userId', oAuthUser._id);
-
-			if (loginClient) {
-				redirectUrl.searchParams.set('loginClient', loginClient);
-			}
-
-			res.redirect(redirectUrl.toString());
-
-			req.session.destroy((err) => {
-				if (err) {
-					console.error('Error destroying session', err);
-				}
-			});
-		},
+		passportOAuthCallback(siteUrl),
 	);
 };
